Update Financial Manager Version 1.py
This is a pretty big update on the module. I have decided to change the class selection input from text to numbers. The way this works is that you have to enter the number 1, 2, or 3 so that you can choose if you want low income, middle class, or high income. I have also added a level of visual drama by adding a loading counter using the time module. This also gives the user time to read the instruction in the shell.

I also have plans to make this project extendable i.e, this project contains a whole package that has retirement planners, mutual fund planners, back savings calculators, etc. By the time of this update, I seem to have gained some extra knowledge on what I want to do with this project.